+++ packages/components-js/tests/e2e/puppeteer/specs/tabs.e2e.ts
@@ -340,13 +340,13 @@ describe('lifecycle', () => {
     const status = await getLifecycleStatus(page);
 
     expect(status.componentDidLoad['p-tabs'], 'componentDidLoad: p-tabs').toBe(1);
-    expect(status.componentDidLoad['p-tabs-bar'], 'componentDidLoad: p-tabs-bar').toBe(1); // includes 6 didLoad calls
+    expect(status.componentDidLoad['p-tabs-bar'], 'componentDidLoad: p-tabs-bar').toBe(1); // includes 4 didLoad calls
     expect(status.componentDidLoad['p-tabs-item'], 'componentDidLoad: p-tabs-item').toBe(3);
 
     expect(status.componentDidUpdate['p-scroller'], 'componentDidUpdate: p-scroller').toBe(0);
     expect(status.componentDidUpdate['p-tabs-bar'], 'componentDidUpdate: p-tabs-bar').toBe(0);
 
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(10);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(8);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(0);
   });
 
@@ -363,7 +363,7 @@ describe('lifecycle', () => {
     expect(status.componentDidUpdate['p-tabs-bar'], 'componentDidUpdate: p-tabs-bar').toBe(1);
     expect(status.componentDidUpdate['p-tabs-item'], 'componentDidUpdate: p-tabs-item').toBe(3);
 
-    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(10);
+    expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(8);
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(5);
   });
 });