+++ packages/components-vue/tests/e2e/specs/partials.e2e.ts
@@ -1,6 +1,6 @@
 import * as jsPartials from '@porsche-design-system/components-js/partials';
 import * as vuePartials from '@porsche-design-system/components-vue/partials';
 
-it('should reexport all partials from components-js', () => {
+it('should reexport partials sub-package from components-js', () => {
   expect(vuePartials).toEqual(jsPartials);
 });