+++ packages/components/src/components/scroller/scroller.tsx
@@ -44,6 +44,7 @@ export class Scroller {
   /** Adapts the color when used on dark background. */
   @Prop() public theme?: Theme = 'light';
 
+  // TODO: Naming is strange? Theme or Scheme
   /** Adapts the background gradient color of prev and next button. */
   @Prop() public gradientColorScheme?: GradientColorTheme = 'default';
 
@@ -108,19 +109,15 @@ export class Scroller {
       const PrefixedTagNames = getPrefixedTagNames(this.host);
       return (
         <div key={direction} class={direction === 'next' ? 'action-next' : 'action-prev'}>
-          <PrefixedTagNames.pButtonPure
-            class="button"
+          <button
             type="button"
             tabIndex={-1}
-            hideLabel={true}
-            size="inherit"
-            icon={direction === 'next' ? 'arrow-head-right' : 'arrow-head-left'}
             onClick={() => this.scrollOnPrevNextClick(direction)}
-            theme={this.theme}
             aria-hidden="true"
+            aria-label={direction}
           >
-            {direction}
-          </PrefixedTagNames.pButtonPure>
+            <PrefixedTagNames.pIcon class="icon" name={direction === 'next' ? 'arrow-head-right' : 'arrow-head-left'} />
+          </button>
         </div>
       );
     };