+++ packages/storefront/src/components/ColorBadge.vue
@@ -52,7 +52,7 @@
 </script>
 
 <style scoped lang="scss">
-  @import '~@porsche-design-system/components-js/utilities/scss';
+  @import '~@porsche-design-system/components-js/styles/scss';
 
   .color-badge {
     width: 10rem;