+++ packages/utilities/projects/utilities/src/js/gradient/gradientLeftStyle.ts
@@ -1,5 +1,0 @@
-import { _gradient } from './gradientShared';
-
-export const gradientLeftStyle = {
-  backgroundColor: `linear-gradient(to left, ${_gradient} 100%);`,
-} as const;