+++ packages/utilities/projects/utilities/src/js/gradient/gradientToBottomStyle.ts
@@ -1,0 +1,5 @@
+import { _gradient } from './gradientShared';
+
+export const gradientToBottomStyle = {
+  background: `linear-gradient(to bottom, ${_gradient} 100%);`,
+} as const;